requirement.txt
alembic==1.2.1
amqp==2.5.2
aniso8601==8.0.0
appnope==0.1.0
astroid==2.3.2
backcall==0.1.0
billiard==3.6.1.0
Blueprints==2.3.0.2
bson==0.5.8
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.0.3
django-cors-middleware==1.5.0
django-extensions==2.2.8
django-mysql==3.3.0
django-oauth-toolkit==1.2.0
django-ratelimit==2.0.0
djangorestframework==3.11.0
dnspython==1.16.0
dominate==2.4.0
ecdsa==0.15
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.9
image==1.5.27
importlib-metadata==1.4.0
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mongo==0.2.0
monotonic==1.5
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pylint-flask-sqlalchemy==0.2.0
pymodm==0.4.2
pymongo==3.9.0b1
pyparsing==2.4.5
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.11.2
python-jose==3.1.0
python-socketio==4.4.0
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.7
visitor==0.1.3
wcwidth==0.1.8
Werkzeug==0.16.0
wordcloud==1.6.0
wrapt==1.11.2
WTForms==2.2.1
zipp==2.1.0
